fedpoly/portal/requirements.txt

django-filter==2.0.0
djangorestframework==3.8.2
Markdown==2.6.11
Pillow==5.2.0
pytz==2018.5
django-debug-toolbar==1.10
